+++ programs/drift/src/ids.rs
@@ -29,63 +29,6 @@ pub mod switchboard_on_demand {
     declare_id!("Aio4gaXjXzJNVLtzwtNVmSqGKpANtXhybbkhtAC94ji2");
 }
 
-pub mod bonk_oracle {
-    use solana_program::declare_id;
-    #[cfg(feature = "mainnet-beta")]
-    declare_id!("8ihFLu5FimgTQ1Unh4dVyEHUGodJ5gJQCrQf4KUVB9bN");
-    #[cfg(not(feature = "mainnet-beta"))]
-    declare_id!("6bquU99ktV1VRiHDr8gMhDFt3kMfhCQo5nfNrg2Urvsn");
-}
-
-pub mod bonk_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("GojbSnJuPdKDT1ZuHuAM5t9oz6bxTo1xhUKpTua2F72p");
-}
-
-pub mod pepe_oracle {
-    use solana_program::declare_id;
-    #[cfg(feature = "mainnet-beta")]
-    declare_id!("FSfxunDmjjbDV2QxpyxFCAPKmYJHSLnLuvQXDLkMzLBm");
-    #[cfg(not(feature = "mainnet-beta"))]
-    declare_id!("Gz9RfgDeAFSsH7BHDGyNTgCik74rjNwsodJpsCizzmkj");
-}
-
-pub mod pepe_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("CLxofhtzvLiErpn25wvUzpZXEqBhuZ6WMEckEraxyuGt");
-}
-
-pub mod wen_oracle {
-    use solana_program::declare_id;
-    #[cfg(feature = "mainnet-beta")]
-    declare_id!("6Uo93N83iF5U9KwC8eQpogx4XptMT4wSKfje7hB1Ufko");
-    #[cfg(not(feature = "mainnet-beta"))]
-    declare_id!("HuobqtT6QaJ8napVARKRxqZN33NqYzQJKLTKKrGy8Bvo");
-}
-
-pub mod wen_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("F47c7aJgYkfKXQ9gzrJaEpsNwUKHprysregTWXrtYLFp");
-}
-
-pub mod mew_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("DKGwCUcwngwmgifGxnme7zVR695LCBGk2pnuksRnbhfD");
-}
-
-pub mod usdc_oracle {
-    use solana_program::declare_id;
-    #[cfg(feature = "mainnet-beta")]
-    declare_id!("Gnt27xtC473ZT2Mw5u8wZ68Z3gULkSTb5DuxJy7eJotD");
-    #[cfg(not(feature = "mainnet-beta"))]
-    declare_id!("5SSkXsEKQepHHAewytPVwdej4epN1nxgLVM84L4KXgy7");
-}
-
-pub mod usdc_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("En8hkHLkRe9d9DraYmBTrus518BvmVH448YcvmrFM6Ce");
-}
-
 pub mod serum_program {
     use solana_program::declare_id;
     #[cfg(feature = "mainnet-beta")]
@@ -120,16 +63,6 @@ pub mod jupiter_mainnet_3 {
 pub mod marinade_mainnet {
     use solana_program::declare_id;
     declare_id!("MarBmsSgKXdrN1egZf5sqe1TMai9K1rChYNDJgjq7aD");
-}
-
-pub mod usdt_oracle {
-    use solana_program::declare_id;
-    declare_id!("3vxLXJqLqF3JG5TCbYycbKWRBbCJQLxQmBGCkyqEEefL");
-}
-
-pub mod usdt_pull_oracle {
-    use solana_program::declare_id;
-    declare_id!("BekJ3P5G3iFeC97sXHuKnUHofCFj9Sbo7uyF2fkKwvit");
 }
 
 pub mod admin_hot_wallet {