+++ programs/drift/src/instructions/admin.rs
@@ -194,9 +194,9 @@ pub fn handle_initialize_spot_market(
             )
         } else {
             validate!(
-                ctx.accounts.spot_market_mint.decimals >= 6,
+                ctx.accounts.spot_market_mint.decimals >= 5,
                 ErrorCode::InvalidSpotMarketInitialization,
-                "Mint decimals must be greater than or equal to 6"
+                "Mint decimals must be greater than or equal to 5"
             )?;
 
             validate!(