+++ programs/drift/src/instructions/user.rs
@@ -357,7 +357,7 @@ pub fn handle_deposit<'c: 'info, 'info>(
     validate!(!user.is_bankrupt(), ErrorCode::UserBankrupt)?;
 
     let mut spot_market = spot_market_map.get_ref_mut(&market_index)?;
-    let oracle_price_data = &oracle_map.get_price_data(&spot_market.oracle)?.clone();
+    let oracle_price_data = *oracle_map.get_price_data(&spot_market.oracle_id())?;
 
     validate!(
         user.pool_id == spot_market.pool_id,
@@ -375,7 +375,7 @@ pub fn handle_deposit<'c: 'info, 'info>(
 
     controller::spot_balance::update_spot_market_cumulative_interest(
         &mut spot_market,
-        Some(oracle_price_data),
+        Some(&oracle_price_data),
         now,
     )?;
 
@@ -533,7 +533,7 @@ pub fn handle_withdraw<'c: 'info, 'info>(
 
     let spot_market_is_reduce_only = {
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
-        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle)?;
+        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle_id())?;
 
         controller::spot_balance::update_spot_market_cumulative_interest(
             spot_market,
@@ -576,7 +576,7 @@ pub fn handle_withdraw<'c: 'info, 'info>(
         };
 
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
-        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle)?;
+        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle_id())?;
 
         user.increment_total_withdraws(
             amount,
@@ -615,7 +615,7 @@ pub fn handle_withdraw<'c: 'info, 'info>(
     user.update_last_active_slot(slot);
 
     let mut spot_market = spot_market_map.get_ref_mut(&market_index)?;
-    let oracle_price = oracle_map.get_price_data(&spot_market.oracle)?.price;
+    let oracle_price = oracle_map.get_price_data(&spot_market.oracle_id())?.price;
 
     let is_borrow = user
         .get_spot_position(market_index)
@@ -724,7 +724,7 @@ pub fn handle_transfer_deposit<'c: 'info, 'info>(
 
     {
         let spot_market = &mut spot_market_map.get_ref_mut(&market_index)?;
-        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle)?;
+        let oracle_price_data = oracle_map.get_price_data(&spot_market.oracle_id())?;
         controller::spot_balance::update_spot_market_cumulative_interest(
             spot_market,
             Some(oracle_price_data),
@@ -734,7 +734,7 @@ pub fn handle_transfer_deposit<'c: 'info, 'info>(
 
     let oracle_price = {
         let spot_market = &spot_market_map.get_ref(&market_index)?;
-        oracle_map.get_price_data(&spot_market.oracle)?.price
+        oracle_map.get_price_data(&spot_market.oracle_id())?.price
     };
 
     {
@@ -3069,7 +3069,7 @@ pub fn handle_begin_swap<'c: 'info, 'info>(
         "begin_swap ended in invalid state"
     )?;
 
-    let in_oracle_data = oracle_map.get_price_data(&in_spot_market.oracle)?;
+    let in_oracle_data = oracle_map.get_price_data(&in_spot_market.oracle_id())?;
     controller::spot_balance::update_spot_market_cumulative_interest(
         &mut in_spot_market,
         Some(in_oracle_data),
@@ -3092,7 +3092,7 @@ pub fn handle_begin_swap<'c: 'info, 'info>(
         "begin_swap ended in invalid state"
     )?;
 
-    let out_oracle_data = oracle_map.get_price_data(&out_spot_market.oracle)?;
+    let out_oracle_data = oracle_map.get_price_data(&out_spot_market.oracle_id())?;
     controller::spot_balance::update_spot_market_cumulative_interest(
         &mut out_spot_market,
         Some(out_oracle_data),
@@ -3222,30 +3222,40 @@ pub fn handle_begin_swap<'c: 'info, 'info>(
                 )?;
             }
         } else {
-            let mut whitelisted_programs = vec![
-                serum_program::id(),
-                AssociatedToken::id(),
-                jupiter_mainnet_3::ID,
-                jupiter_mainnet_4::ID,
-                jupiter_mainnet_6::ID,
-            ];
-            if !delegate_is_signer {
-                whitelisted_programs.push(Token::id());
-                whitelisted_programs.push(Token2022::id());
-                whitelisted_programs.push(marinade_mainnet::ID);
-            }
-            validate!(
-                whitelisted_programs.contains(&ix.program_id),
-                ErrorCode::InvalidSwap,
-                "only allowed to pass in ixs to token, openbook, and Jupiter v3/v4/v6 programs"
-            )?;
-
-            for meta in ix.accounts.iter() {
+            if found_end {
+                for meta in ix.accounts.iter() {
+                    validate!(
+                        meta.is_writable == false,
+                        ErrorCode::InvalidSwap,
+                        "instructions after swap end must not have writable accounts"
+                    )?;
+                }
+            } else {  
+                let mut whitelisted_programs = vec![
+                    serum_program::id(),
+                    AssociatedToken::id(),
+                    jupiter_mainnet_3::ID,
+                    jupiter_mainnet_4::ID,
+                    jupiter_mainnet_6::ID,
+                ];
+                if !delegate_is_signer {
+                    whitelisted_programs.push(Token::id());
+                    whitelisted_programs.push(Token2022::id());
+                    whitelisted_programs.push(marinade_mainnet::ID);
+                }
                 validate!(
-                    meta.pubkey != crate::id(),
+                    whitelisted_programs.contains(&ix.program_id),
                     ErrorCode::InvalidSwap,
-                    "instructions between begin and end must not be drift instructions"
+                    "only allowed to pass in ixs to token, openbook, and Jupiter v3/v4/v6 programs"
                 )?;
+
+                for meta in ix.accounts.iter() {
+                    validate!(
+                        meta.pubkey != crate::id(),
+                        ErrorCode::InvalidSwap,
+                        "instructions between begin and end must not be drift instructions"
+                    )?;
+                }
             }
         }
 
@@ -3326,7 +3336,7 @@ pub fn handle_end_swap<'c: 'info, 'info>(
         "the in_spot_market must have a flash loan amount set"
     )?;
 
-    let in_oracle_data = oracle_map.get_price_data(&in_spot_market.oracle)?;
+    let in_oracle_data = oracle_map.get_price_data(&in_spot_market.oracle_id())?;
     let in_oracle_price = in_oracle_data.price;
 
     let mut out_spot_market = spot_market_map.get_ref_mut(&out_market_index)?;
@@ -3338,7 +3348,7 @@ pub fn handle_end_swap<'c: 'info, 'info>(
         out_market_index
     )?;
 
-    let out_oracle_data = oracle_map.get_price_data(&out_spot_market.oracle)?;
+    let out_oracle_data = oracle_map.get_price_data(&out_spot_market.oracle_id())?;
     let out_oracle_price = out_oracle_data.price;
 
     let in_vault = &mut ctx.accounts.in_spot_market_vault;