+++ tests/placeAndMakeSwiftPerpBankrun.ts
@@ -32,6 +32,7 @@ import {
 	ANCHOR_TEST_SWIFT_ID,
 	SwiftOrderRecord,
 	getSwiftUserAccountPublicKey,
+	digest,
 } from '../sdk/src';
 
 import {
@@ -244,7 +245,7 @@ describe('place and make swift order', () => {
 			makerDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = makerDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);
 
@@ -415,7 +416,7 @@ describe('place and make swift order', () => {
 			swiftDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = swiftDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);
 
@@ -534,7 +535,7 @@ describe('place and make swift order', () => {
 			takerDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = takerDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);
 
@@ -613,7 +614,7 @@ describe('place and make swift order', () => {
 			takerDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = takerDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);
 
@@ -711,7 +712,7 @@ describe('place and make swift order', () => {
 			makerDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = makerDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);
 
@@ -788,7 +789,7 @@ describe('place and make swift order', () => {
 			makerDriftClient.encodeSwiftServerMessage(swiftServerMessage);
 
 		const swiftSignature = makerDriftClient.signMessage(
-			Uint8Array.from(encodedSwiftServerMessage),
+			digest(encodedSwiftServerMessage),
 			swiftKeypair
 		);